Streamlined Winner-checking function
A little though reveals that you only need to check four directions to verify if someone has won (North, North-East, East, and South-East). The reason is that, for example, if there is some vertical line then there must always be a bottom element of that line. This means that you will always catch a vertical line if you only check the north direction. A similar argument can be used on the other three directions

Further, to make access a little more intuitivie, the checking function directions has been redone in terms of board directions rather than matrix directions

this code has been tested with N = 4 and appears to work.